+++ tests/Unit/Private/Invoke-SetupAction.Tests.ps1
@@ -149,7 +149,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Install''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -858,7 +858,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Upgrade''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -1040,7 +1040,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''InstallFailoverCluster''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -1531,7 +1531,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''PrepareFailoverCluster''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -1836,7 +1836,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''CompleteFailoverCluster''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -2140,7 +2140,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''AddNode''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -2409,7 +2409,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''RemoveNode''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -2523,7 +2523,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When using a ''ConfigurationFile''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
 
             Mock -CommandName Test-Path -ParameterFilter {
@@ -2677,7 +2677,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''RebuildDatabase''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -2838,7 +2838,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''EditionUpgrade''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -2960,7 +2960,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Repair''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -3137,7 +3137,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Upgrade''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -3303,7 +3303,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''PrepareImage''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -3479,7 +3479,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''CompleteImage''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -3823,7 +3823,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Uninstall''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -3897,7 +3897,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When setup action is ''Install'' for installing Azure Arc Agent (parameter set InstallAzureArcAgent)' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'
@@ -4043,7 +4043,7 @@ Describe 'Invoke-SetupAction' -Tag 'Public' {
 
     Context 'When parameter set is ''InstallRole''' {
         BeforeAll {
-            Mock -CommandName Assert-InstallSqlServerProperties
+            Mock -CommandName Assert-SetupActionProperties
             Mock -CommandName Assert-ElevatedUser
             Mock -CommandName Test-Path -ParameterFilter {
                 $Path -match 'setup\.exe'